fix nextStats
This never really worked correctly as it prepended all items everytime
with the new item count, so the list grew larger and larger. It didn't
break however because the first item found for a uuid always was the one
with the current count.

Thanks to johthor for some functional help :)